problems/src/array_matrix/_48_RotateImage.java
package array_matrix;

//https://leetcode.com/problems/rotate-image/

public class _48_RotateImage {

    public void rotate(int[][] matrix) {
        int n = matrix.length;
        // transpose
        for (int i = 1; i < n; i++) {
            for (int j = 0; j < i; j++) {
                int tmp = matrix[i][j];
                matrix[i][j] = matrix[j][i];
                matrix[j][i] = tmp;
            }
        }

        // vertical mirror
        for (int i = 0; i < n; i++) {
            for (int j = 0; j < n / 2; j++) {
                int tmp = matrix[i][j];
                matrix[i][j] = matrix[i][n - 1 - j];
                matrix[i][n - 1 - j] = tmp;
            }
        }
    }
}
